refactor: 迁移 language 相关文件到 src/language/ (Phase 4b)
目录重组第二步：将 CSM 语言功能文件迁移到 src/language/。
- hoverData/ → language/hoverData/
- logFold/ → language/logFold/
- csmlogHoverProvider.ts → language/csmlogHoverProvider.ts
- csmlogDocumentSymbolProvider.ts → language/csmlogDocumentSymbolProvider.ts
- lvcsmDocumentSymbolProvider.ts → language/lvcsmDocumentSymbolProvider.ts
- fileDecorationProvider.ts → language/fileDecorationProvider.ts
- hoverData.ts → language/hoverData.ts

更新 extension.ts 及所有测试文件的导入路径。
旧路径保留 re-export 兼容层。352 测试全部通过。
