+++ src/hoverData/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @deprecated 请使用 '../language/hoverData' 代替。
+ * 此文件保留以便平滑迁移，将在后续版本中移除。
+ */
+export * from '../language/hoverData';
