+++ src/language/fileDecorationProvider.ts
