+++ src/language/hoverData/config.ts
