+++ src/language/hoverData/operators.ts
