+++ src/language/hoverData/translations.ts
