+++ src/language/logFold/foldingProvider.ts
