+++ src/language/logFold/types.ts
