+++ src/test/hoverData.test.ts
@@ -33,7 +33,7 @@ function makeDoc(lines: string[], uriOverride?: string): FakeDocument {
 
 // Load the compiled module (vscode is already intercepted by setup.js)
 const hoverData = require(
-    path.resolve(__dirname, '../hoverData')
+    path.resolve(__dirname, '../language/hoverData')
 ) as {
     clearAnchorCache: (uri: string) => void;
     provideContentHover: (doc: FakeDocument, pos: any) => any;
