+++ src/test/logFold/detector.test.ts
@@ -5,9 +5,9 @@
  */
 
 import * as assert from 'assert';
-import { detectRepeatRegions } from '../../logFold/detector';
-import { normalizeLine } from '../../logFold/normalizer';
-import { DEFAULT_FOLD_OPTIONS, FoldOptions, LineSignature } from '../../logFold/types';
+import { detectRepeatRegions } from '../../language/logFold/detector';
+import { normalizeLine } from '../../language/logFold/normalizer';
+import { DEFAULT_FOLD_OPTIONS, FoldOptions, LineSignature } from '../../language/logFold/types';
 
 // ---------------------------------------------------------------------------
 // 辅助
